+++ src/math/mod.rs
@@ -5,6 +5,7 @@ mod integer;
 pub use crate::simd::SimdAnyInt;
 pub use integer::*;
 
+pub(crate) use cast::*;
 pub(crate) use float::*;
 
 #[cfg(any(feature = "fuzzing", test))]